+++ docs/csharp/how-to/snippets/strings/Program.cs
@@ -1,0 +1,17 @@
+﻿using HowToStrings;
+
+Console.WriteLine("============================  String.Split examples =================================================");
+Console.WriteLine();
+ParseStringsUsingSplit.Examples();
+Console.WriteLine("============================  String concatenation examples =================================================");
+Console.WriteLine();
+Concatenate.Examples();
+Console.WriteLine("============================  String Searching examples =================================================");
+Console.WriteLine();
+SearchStrings.Examples();
+Console.WriteLine("============================  Modify string examples =================================================");
+Console.WriteLine();
+ModifyStrings.Examples();
+Console.WriteLine("============================  Compare string examples =================================================");
+Console.WriteLine();
+CompareStrings.Examples();
